+++ src/main/java/net/loganford/noideaengine/state/entity/Entity2D.java
@@ -75,7 +75,7 @@ public abstract class Entity2D<G extends Game, S extends Scene> extends Abstract
     }
 
     @Override
-    public void step(G game, S scene, long delta) {
+    public void step(G game, S scene, float delta) {
         sprite.step(delta);
     }
