+++ src/main/java/net/loganford/noideaengine/state/loading/LoadingScreen.java
@@ -45,7 +45,7 @@ public abstract class LoadingScreen extends GameState {
     }
 
     @Override
-    public void step(Game game, long delta) {
+    public void step(Game game, float delta) {
         super.step(game, delta);
 
         {
